+++ cmake/SearchForStuff.cmake
@@ -114,7 +114,7 @@ if (PKG_CONFIG_FOUND)
 
   #################################################
   # Find DART
-  find_package(DARTCore 4.1 QUIET)
+  find_package(DARTCore 4.3 QUIET)
   if (DARTCore_FOUND)
     message (STATUS "Looking for DARTCore - found")
     set (HAVE_DART TRUE)
@@ -190,7 +190,16 @@ if (PKG_CONFIG_FOUND)
   # Find OGRE
   execute_process(COMMAND pkg-config --modversion OGRE
                   OUTPUT_VARIABLE OGRE_VERSION)
-  string(REPLACE "\n" "" OGRE_VERSION ${OGRE_VERSION})
+
+  string (REGEX REPLACE "^([0-9]+).*" "\\1"
+    OGRE_MAJOR_VERSION "${OGRE_VERSION}")
+  string (REGEX REPLACE "^[0-9]+\\.([0-9]+).*" "\\1"
+    OGRE_MINOR_VERSION "${OGRE_VERSION}")
+  string (REGEX REPLACE "^[0-9]+\\.[0-9]+\\.([0-9]+).*" "\\1"
+    OGRE_PATCH_VERSION ${OGRE_VERSION})
+
+  set(OGRE_VERSION
+    ${OGRE_MAJOR_VERSION}.${OGRE_MINOR_VERSION}.${OGRE_PATCH_VERSION})
 
   pkg_check_modules(OGRE-RTShaderSystem
                     OGRE-RTShaderSystem>=${MIN_OGRE_VERSION})