+++ gazebo/gui/model/ModelCreator.cc
@@ -15,23 +15,20 @@
  *
  */
 
+#include <boost/thread/recursive_mutex.hpp>
 #include <boost/filesystem.hpp>
 #include <sstream>
 #include <string>
 
-#include "gazebo/common/KeyEvent.hh"
 #include "gazebo/common/Exception.hh"
 
 #include "gazebo/rendering/UserCamera.hh"
-#include "gazebo/rendering/Visual.hh"
 #include "gazebo/rendering/Scene.hh"
 
 #include "gazebo/math/Quaternion.hh"
 
 #include "gazebo/transport/Publisher.hh"
 #include "gazebo/transport/Node.hh"
-
-#include "gazebo/physics/Inertial.hh"
 
 #include "gazebo/gui/Actions.hh"
 #include "gazebo/gui/KeyEventHandler.hh"
@@ -44,6 +41,7 @@
 #include "gazebo/gui/SaveDialog.hh"
 
 #include "gazebo/gui/model/ModelData.hh"
+#include "gazebo/gui/model/LinkInspector.hh"
 #include "gazebo/gui/model/JointMaker.hh"
 #include "gazebo/gui/model/ModelEditorEvents.hh"
 #include "gazebo/gui/model/ModelCreator.hh"
@@ -60,15 +58,13 @@ ModelCreator::ModelCreator()
   this->active = false;
 
   this->modelTemplateSDF.reset(new sdf::SDF);
-  this->modelTemplateSDF->SetFromString(this->GetTemplateSDFString());
+  this->modelTemplateSDF->SetFromString(ModelData::GetTemplateSDFString());
+
+  this->updateMutex = new boost::recursive_mutex();
 
   this->manipMode = "";
-  this->boxCounter = 0;
-  this->cylinderCounter = 0;
-  this->sphereCounter = 0;
+  this->partCounter = 0;
   this->modelCounter = 0;
-
-  this->editTransparency = 0.4;
 
   this->node = transport::NodePtr(new transport::Node());
   this->node->Init();
@@ -78,6 +74,10 @@ ModelCreator::ModelCreator()
   this->jointMaker = new JointMaker();
 
   connect(g_editModelAct, SIGNAL(toggled(bool)), this, SLOT(OnEdit(bool)));
+  this->inspectAct = new QAction(tr("Open Link Inspector"), this);
+  connect(this->inspectAct, SIGNAL(triggered()), this,
+      SLOT(OnOpenInspector()));
+
   connect(g_deleteAct, SIGNAL(DeleteSignal(const std::string &)), this,
           SLOT(OnDelete(const std::string &)));
 
@@ -118,6 +118,10 @@ ModelCreator::ModelCreator()
        boost::bind(&ModelCreator::OnSetSelectedEntity, this, _1, _2)));
 
   this->connections.push_back(
+      gui::Events::ConnectScaleEntity(
+      boost::bind(&ModelCreator::OnEntityScaleChanged, this, _1, _2)));
+
+  this->connections.push_back(
       event::Events::ConnectPreRender(
         boost::bind(&ModelCreator::Update, this)));
 
@@ -142,6 +146,8 @@ ModelCreator::~ModelCreator()
   this->requestPub.reset();
   this->makerPub.reset();
   this->connections.clear();
+
+  delete this->updateMutex;
 
   delete jointMaker;
 }
@@ -401,7 +407,7 @@ std::string ModelCreator::AddBox(const math::Vector3 &_size,
   }
 
   std::ostringstream linkNameStream;
-  linkNameStream << "unit_box_" << this->boxCounter++;
+  linkNameStream << "part_" << this->partCounter++;
   std::string linkName = linkNameStream.str();
 
   rendering::VisualPtr linkVisual(new rendering::Visual(linkName,
@@ -421,12 +427,12 @@ std::string ModelCreator::AddBox(const math::Vector3 &_size,
 
   visVisual->Load(visualElem);
 
-  linkVisual->SetTransparency(this->editTransparency);
+  linkVisual->SetTransparency(ModelData::GetEditTransparency());
   linkVisual->SetPose(_pose);
   if (_pose == math::Pose::Zero)
   {
     linkVisual->SetPosition(math::Vector3(_pose.pos.x, _pose.pos.y,
-    _pose.pos.z + _size.z/2));
+    _pose.pos.z + _size.z*0.5));
   }
 
   this->CreatePart(visVisual);
@@ -443,7 +449,7 @@ std::string ModelCreator::AddSphere(double _radius,
     this->Reset();
 
   std::ostringstream linkNameStream;
-  linkNameStream << "unit_sphere_" << this->sphereCounter++;
+  linkNameStream << "part_" << this->partCounter++;
   std::string linkName = linkNameStream.str();
 
   rendering::VisualPtr linkVisual(new rendering::Visual(
@@ -463,7 +469,7 @@ std::string ModelCreator::AddSphere(double _radius,
 
   visVisual->Load(visualElem);
 
-  linkVisual->SetTransparency(this->editTransparency);
+  linkVisual->SetTransparency(ModelData::GetEditTransparency());
   linkVisual->SetPose(_pose);
   if (_pose == math::Pose::Zero)
   {
@@ -472,6 +478,7 @@ std::string ModelCreator::AddSphere(double _radius,
   }
 
   this->CreatePart(visVisual);
+
   this->mouseVisual = linkVisual;
 
   return linkName;
@@ -485,7 +492,7 @@ std::string ModelCreator::AddCylinder(double _radius, double _length,
     this->Reset();
 
   std::ostringstream linkNameStream;
-  linkNameStream << "unit_cylinder_" << this->cylinderCounter++;
+  linkNameStream << "part_" << this->partCounter++;
   std::string linkName = linkNameStream.str();
 
   rendering::VisualPtr linkVisual(new rendering::Visual(
@@ -507,15 +514,16 @@ std::string ModelCreator::AddCylinder(double _radius, double _length,
 
   visVisual->Load(visualElem);
 
-  linkVisual->SetTransparency(this->editTransparency);
+  linkVisual->SetTransparency(ModelData::GetEditTransparency());
   linkVisual->SetPose(_pose);
   if (_pose == math::Pose::Zero)
   {
     linkVisual->SetPosition(math::Vector3(_pose.pos.x, _pose.pos.y,
-    _pose.pos.z + _length/2));
+    _pose.pos.z + _length*0.5));
   }
 
   this->CreatePart(visVisual);
+
   this->mouseVisual = linkVisual;
 
   return linkName;
@@ -531,11 +539,11 @@ std::string ModelCreator::AddCustom(const std::string &_path,
   std::string path = _path;
 
   std::ostringstream linkNameStream;
-  linkNameStream << "custom_" << this->customCounter++;
+  linkNameStream << "part_" << this->partCounter++;
   std::string linkName = linkNameStream.str();
 
-  rendering::VisualPtr linkVisual(new rendering::Visual(this->previewName +
-      "::" + linkName, this->previewVisual));
+  rendering::VisualPtr linkVisual(new rendering::Visual(
+      linkName, this->previewVisual));
   linkVisual->Load();
 
   std::ostringstream visualName;
@@ -552,15 +560,16 @@ std::string ModelCreator::AddCustom(const std::string &_path,
   meshElem->GetElement("uri")->Set(path);
   visVisual->Load(visualElem);
 
-  linkVisual->SetTransparency(this->editTransparency);
+  linkVisual->SetTransparency(ModelData::GetEditTransparency());
   linkVisual->SetPose(_pose);
   if (_pose == math::Pose::Zero)
   {
     linkVisual->SetPosition(math::Vector3(_pose.pos.x, _pose.pos.y,
-    _pose.pos.z + _scale.z/2));
+    _pose.pos.z + _scale.z*0.5));
   }
 
   this->CreatePart(visVisual);
+
   this->mouseVisual = linkVisual;
 
   return linkName;
@@ -569,22 +578,32 @@ std::string ModelCreator::AddCustom(const std::string &_path,
 /////////////////////////////////////////////////
 void ModelCreator::CreatePart(const rendering::VisualPtr &_visual)
 {
-  PartData *part = new PartData;
-
+  PartData *part = new PartData();
   part->partVisual = _visual->GetParent();
-  part->visuals.push_back(_visual);
-
-  part->name = part->partVisual->GetName();
-  part->pose = part->partVisual->GetWorldPose();
   part->scale = part->partVisual->GetScale();
-  part->gravity = true;
-  part->selfCollide = false;
-  part->kinematic = false;
+  part->AddVisual(_visual);
 
-  this->allParts[part->name] = part;
+  // create collision with identical geometry
+  rendering::VisualPtr collisionVis =
+      _visual->Clone(part->partVisual->GetName() + "_collision",
+      part->partVisual);
 
-  rendering::ScenePtr scene = part->partVisual->GetScene();
-  scene->AddVisual(part->partVisual);
+  // orange
+  collisionVis->SetAmbient(common::Color(1.0, 0.5, 0.05));
+  collisionVis->SetDiffuse(common::Color(1.0, 0.5, 0.05));
+  collisionVis->SetSpecular(common::Color(0.5, 0.5, 0.5));
+  collisionVis->SetTransparency(
+      math::clamp(ModelData::GetEditTransparency() * 2.0, 0.0, 0.8));
+  // fix for transparency alpha compositing
+  Ogre::MovableObject *colObj = collisionVis->GetSceneNode()->
+      getAttachedObject(0);
+  colObj->setRenderQueueGroup(colObj->getRenderQueueGroup()+1);
+  part->AddCollision(collisionVis);
+
+  std::string partName = part->partVisual->GetName();
+  part->SetName(partName);
+  part->SetPose(part->partVisual->GetWorldPose());
+  this->allParts[partName] = part;
 
   this->ModelChanged();
 }
@@ -606,14 +625,22 @@ void ModelCreator::RemovePart(const std::string &_partName)
     return;
 
   rendering::ScenePtr scene = part->partVisual->GetScene();
-  for (unsigned int i = 0; i < part->visuals.size(); ++i)
+  for (auto &it : part->visuals)
   {
-    rendering::VisualPtr vis = part->visuals[i];
+    rendering::VisualPtr vis = it.first;
     scene->RemoveVisual(vis);
   }
   scene->RemoveVisual(part->partVisual);
-  part->visuals.clear();
+  for (auto &colIt : part->collisions)
+  {
+    rendering::VisualPtr vis = colIt.first;
+    scene->RemoveVisual(vis);
+  }
+
+  scene->RemoveVisual(part->partVisual);
+
   part->partVisual.reset();
+  delete part->inspector;
 
   this->allParts.erase(_partName);
   this->ModelChanged();
@@ -640,7 +667,7 @@ void ModelCreator::Reset()
   this->isStatic = false;
   this->autoDisable = true;
 
-  while (this->allParts.size() > 0)
+  while (!this->allParts.empty())
     this->RemovePart(this->allParts.begin()->first);
   this->allParts.clear();
 
@@ -729,36 +756,6 @@ void ModelCreator::CreateTheEntity()
 
   msg.set_sdf(this->modelSDF->ToString());
   this->makerPub->Publish(msg);
-}
-
-/////////////////////////////////////////////////
-std::string ModelCreator::GetTemplateSDFString()
-{
-  std::ostringstream newModelStr;
-  newModelStr << "<sdf version ='" << SDF_VERSION << "'>"
-    << "<model name='template_model'>"
-    << "<pose>0 0 0.0 0 0 0</pose>"
-    << "<link name ='link'>"
-    <<   "<visual name ='visual'>"
-    <<     "<pose>0 0 0.0 0 0 0</pose>"
-    <<     "<geometry>"
-    <<       "<box>"
-    <<         "<size>1.0 1.0 1.0</size>"
-    <<       "</box>"
-    <<     "</geometry>"
-    <<     "<material>"
-    <<       "<script>"
-    <<         "<uri>file://media/materials/scripts/gazebo.material</uri>"
-    <<         "<name>Gazebo/Grey</name>"
-    <<       "</script>"
-    <<     "</material>"
-    <<   "</visual>"
-    << "</link>"
-    << "<static>true</static>"
-    << "</model>"
-    << "</sdf>";
-
-  return newModelStr.str();
 }
 
 /////////////////////////////////////////////////
@@ -923,12 +920,15 @@ bool ModelCreator::OnMouseRelease(const common::MouseEvent &_event)
 
   if (this->mouseVisual)
   {
+    if (_event.button == common::MouseEvent::RIGHT)
+      return true;
+
     // set the part data pose
     if (this->allParts.find(this->mouseVisual->GetName()) !=
         this->allParts.end())
     {
       PartData *part = this->allParts[this->mouseVisual->GetName()];
-      part->pose = this->mouseVisual->GetWorldPose();
+      part->SetPose(this->mouseVisual->GetWorldPose());
     }
 
     // reset and return
@@ -949,6 +949,16 @@ bool ModelCreator::OnMouseRelease(const common::MouseEvent &_event)
       // Handle snap from GLWidget
       if (g_snapAct->isChecked())
         return false;
+
+      // trigger part inspector on right click
+      if (_event.button == common::MouseEvent::RIGHT)
+      {
+        this->inspectVis = vis->GetParent();
+        QMenu menu;
+        menu.addAction(this->inspectAct);
+        menu.exec(QCursor::pos());
+        return true;
+      }
 
       // Not in multi-selection mode.
       if (!(QApplication::keyboardModifiers() & Qt::ControlModifier))
@@ -1066,11 +1076,28 @@ bool ModelCreator::OnMouseDoubleClick(const common::MouseEvent &_event)
   if (this->allParts.find(vis->GetParent()->GetName()) !=
       this->allParts.end())
   {
-    // TODO open inspector.
+    this->OpenInspector(vis->GetParent()->GetName());
     return true;
   }
 
   return false;
+}
+
+/////////////////////////////////////////////////
+void ModelCreator::OnOpenInspector()
+{
+  this->OpenInspector(this->inspectVis->GetName());
+  this->inspectVis.reset();
+}
+
+/////////////////////////////////////////////////
+void ModelCreator::OpenInspector(const std::string &_name)
+{
+  PartData *part = this->allParts[_name];
+  part->SetPose(part->partVisual->GetWorldPose());
+  part->UpdateConfig();
+  part->inspector->move(QCursor::pos());
+  part->inspector->show();
 }
 
 /////////////////////////////////////////////////
@@ -1099,8 +1126,7 @@ void ModelCreator::OnPaste()
   }
 
   // For now, only copy the last selected model
-  boost::unordered_map<std::string, PartData *>::iterator it =
-    this->allParts.find(this->copiedPartNames.back());
+  auto it = this->allParts.find(this->copiedPartNames.back());
   if (it != this->allParts.end())
   {
     PartData *copiedPart = it->second;
@@ -1110,7 +1136,7 @@ void ModelCreator::OnPaste()
     this->Stop();
     this->DeselectAll();
 
-    std::string linkName = copiedPart->name + "_clone";
+    std::string linkName = copiedPart->GetName() + "_clone";
 
     if (!this->previewVisual)
     {
@@ -1138,11 +1164,10 @@ void ModelCreator::OnPaste()
     }
     else
     {
-      rendering::VisualPtr copiedVisual = copiedPart->visuals.back();
+      rendering::VisualPtr copiedVisual = copiedPart->visuals.rbegin()->first;
       visVisual = copiedVisual->Clone(visualName.str(), linkVisual);
       clonePose = copiedVisual->GetWorldPose();
       cloneScale = copiedVisual->GetParent()->GetScale();
-      visVisual->SetScale(math::Vector3::One);
     }
 
     rendering::UserCameraPtr userCamera = gui::get_active_camera();
@@ -1157,7 +1182,7 @@ void ModelCreator::OnPaste()
 
     linkVisual->SetScale(cloneScale);
     linkVisual->SetWorldPose(clonePose);
-    linkVisual->SetTransparency(this->editTransparency);
+    linkVisual->SetTransparency(ModelData::GetEditTransparency());
 
     this->addPartType = PART_CUSTOM;
     this->CreatePart(visVisual);
@@ -1178,103 +1203,59 @@ void ModelCreator::GenerateSDF()
   sdf::ElementPtr linkElem;
 
   this->modelSDF.reset(new sdf::SDF);
-  this->modelSDF->SetFromString(this->GetTemplateSDFString());
+  this->modelSDF->SetFromString(ModelData::GetTemplateSDFString());
 
   modelElem = this->modelSDF->root->GetElement("model");
 
   linkElem = modelElem->GetElement("link");
   sdf::ElementPtr templateLinkElem = linkElem->Clone();
-  sdf::ElementPtr templateVisualElem = templateLinkElem->GetElement(
-      "visual")->Clone();
-  sdf::ElementPtr templateCollisionElem = templateLinkElem->GetElement(
-      "collision")->Clone();
   modelElem->ClearElements();
   std::stringstream visualNameStream;
   std::stringstream collisionNameStream;
 
   modelElem->GetAttribute("name")->Set(this->folderName);
 
-  boost::unordered_map<std::string, PartData *>::iterator partsIt;
-
   // set center of all parts to be origin
   math::Vector3 mid;
-  for (partsIt = this->allParts.begin(); partsIt != this->allParts.end();
-       ++partsIt)
+  for (auto &partsIt : this->allParts)
   {
-    PartData *part = partsIt->second;
-    mid += part->partVisual->GetWorldPose().pos;
+    PartData *part = partsIt.second;
+    mid += part->GetPose().pos;
   }
   mid /= this->allParts.size();
   this->origin.pos = mid;
   modelElem->GetElement("pose")->Set(this->origin);
 
   // loop through all parts and generate sdf
-  for (partsIt = this->allParts.begin(); partsIt != this->allParts.end();
-       ++partsIt)
+  for (auto &partsIt : this->allParts)
   {
     visualNameStream.str("");
     collisionNameStream.str("");
 
-    PartData *part = partsIt->second;
-    sdf::ElementPtr newLinkElem = templateLinkElem->Clone();
-    newLinkElem->ClearElements();
-    newLinkElem->GetAttribute("name")->Set(part->name);
+    PartData *part = partsIt.second;
+    part->UpdateConfig();
+
+    sdf::ElementPtr newLinkElem = part->partSDF->Clone();
     newLinkElem->GetElement("pose")->Set(part->partVisual->GetWorldPose()
         - this->origin);
-    newLinkElem->GetElement("gravity")->Set(part->gravity);
-    newLinkElem->GetElement("self_collide")->Set(part->selfCollide);
-    newLinkElem->GetElement("kinematic")->Set(part->kinematic);
 
     modelElem->InsertElement(newLinkElem);
 
-    for (unsigned int i = 0; i < part->visuals.size(); ++i)
+    // visuals
+    for (auto const &it : part->visuals)
     {
-      sdf::ElementPtr visualElem = templateVisualElem->Clone();
-      sdf::ElementPtr collisionElem = templateCollisionElem->Clone();
-
-      rendering::VisualPtr visual = part->visuals[i];
-
-      visualElem->GetAttribute("name")->Set(visual->GetName());
-      collisionElem->GetAttribute("name")->Set(
-          visual->GetParent()->GetName() + "_collision");
-      visualElem->GetElement("pose")->Set(visual->GetPose());
-      collisionElem->GetElement("pose")->Set(visual->GetPose());
-
-      sdf::ElementPtr geomElem =  visualElem->GetElement("geometry");
-      geomElem->ClearElements();
-
-      math::Vector3 scale = visual->GetParent()->GetScale();
-      if (visual->GetParent()->GetName().find("unit_box") != std::string::npos)
-      {
-        sdf::ElementPtr boxElem = geomElem->AddElement("box");
-        (boxElem->GetElement("size"))->Set(scale);
-      }
-      else if (visual->GetParent()->GetName().find("unit_cylinder")
-         != std::string::npos)
-      {
-        sdf::ElementPtr cylinderElem = geomElem->AddElement("cylinder");
-        (cylinderElem->GetElement("radius"))->Set(scale.x/2.0);
-        (cylinderElem->GetElement("length"))->Set(scale.z);
-      }
-      else if (visual->GetParent()->GetName().find("unit_sphere")
-          != std::string::npos)
-      {
-        sdf::ElementPtr sphereElem = geomElem->AddElement("sphere");
-        (sphereElem->GetElement("radius"))->Set(scale.x/2.0);
-      }
-      else if (visual->GetParent()->GetName().find("custom")
-          != std::string::npos)
-      {
-        sdf::ElementPtr customElem = geomElem->AddElement("mesh");
-        (customElem->GetElement("scale"))->Set(scale);
-        (customElem->GetElement("uri"))->Set(visual->GetMeshName());
-      }
-      sdf::ElementPtr geomElemClone = geomElem->Clone();
-      geomElem =  collisionElem->GetElement("geometry");
-      geomElem->ClearElements();
-      geomElem->InsertElement(geomElemClone->GetFirstElement());
-
+      rendering::VisualPtr visual = it.first;
+      msgs::Visual visualMsg = it.second;
+      sdf::ElementPtr visualElem = visual->GetSDF()->Clone();
+      visualElem->GetElement("transparency")->Set<double>(
+          visualMsg.transparency());
       newLinkElem->InsertElement(visualElem);
+    }
+
+    // collisions
+    for (auto const &colIt : part->collisions)
+    {
+      sdf::ElementPtr collisionElem = msgs::CollisionToSDF(colIt.second);
       newLinkElem->InsertElement(collisionElem);
     }
   }
@@ -1364,18 +1345,37 @@ void ModelCreator::ModelChanged()
 /////////////////////////////////////////////////
 void ModelCreator::Update()
 {
+  boost::recursive_mutex::scoped_lock lock(*this->updateMutex);
   // Check if any parts have been moved or resized and trigger ModelChanged
-  boost::unordered_map<std::string, PartData *>::iterator partsIt;
-  for (partsIt = this->allParts.begin(); partsIt != this->allParts.end();
-       ++partsIt)
+  for (auto &partsIt : this->allParts)
   {
-    PartData *part = partsIt->second;
-    if (part->pose != part->partVisual->GetWorldPose() ||
-        part->scale != part->partVisual->GetScale())
+    PartData *part = partsIt.second;
+    if (part->GetPose() != part->partVisual->GetWorldPose())
     {
-      part->pose = part->partVisual->GetWorldPose();
-      part->scale = part->partVisual->GetScale();
+      part->SetPose(part->partVisual->GetWorldPose());
       this->ModelChanged();
+    }
+    for (auto scaleIt : this->partScaleUpdate)
+    {
+      if (part->partVisual->GetName() == scaleIt.first)
+        part->SetScale(scaleIt.second);
+    }
+    this->partScaleUpdate.clear();
+  }
+}
+
+/////////////////////////////////////////////////
+void ModelCreator::OnEntityScaleChanged(const std::string &_name,
+  const math::Vector3 &_scale)
+{
+  boost::recursive_mutex::scoped_lock lock(*this->updateMutex);
+  for (auto partsIt : this->allParts)
+  {
+    if (_name == partsIt.first ||
+        _name.find(partsIt.first) != std::string::npos)
+    {
+      this->partScaleUpdate[partsIt.first] = _scale;
+      break;
     }
   }
 }