+++ gazebo/physics/World.cc
@@ -718,11 +718,9 @@ void World::Update()
       boost::recursive_mutex::scoped_lock lock(
         *this->dataPtr->physicsEngine->GetPhysicsUpdateMutex());
 
-      for (std::list<Entity*>::iterator iter =
-          this->dataPtr->dirtyPoses.begin();
-          iter != this->dataPtr->dirtyPoses.end(); ++iter)
+      for (auto &dirtyEntity : this->dataPtr->dirtyPoses)
       {
-        (*iter)->SetWorldPose((*iter)->GetDirtyPose(), false);
+        dirtyEntity->SetWorldPose(dirtyEntity->GetDirtyPose(), false);
       }
 
       this->dataPtr->dirtyPoses.clear();
@@ -793,10 +791,9 @@ void World::ClearModels()
   this->dataPtr->publishModelPoses.clear();
 
   // Remove all models
-  for (Model_V::iterator iter = this->dataPtr->models.begin();
-       iter != this->dataPtr->models.end(); ++iter)
+  for (auto &model : this->dataPtr->models)
   {
-    this->dataPtr->rootElement->RemoveChild((*iter)->GetId());
+    this->dataPtr->rootElement->RemoveChild(model->GetId());
   }
   this->dataPtr->models.clear();
 
@@ -1013,11 +1010,9 @@ void World::Reset()
 
     this->ResetTime();
     this->ResetEntities(Base::BASE);
-    for (std::vector<WorldPluginPtr>::iterator iter =
-        this->dataPtr->plugins.begin();
-        iter != this->dataPtr->plugins.end(); ++iter)
+    for (auto &plugin : this->dataPtr->plugins)
     {
-      (*iter)->Reset();
+      plugin->Reset();
     }
     this->dataPtr->physicsEngine->Reset();
 
@@ -1296,13 +1291,12 @@ void World::LoadPlugin(const std::string &_filename,
 //////////////////////////////////////////////////
 void World::RemovePlugin(const std::string &_name)
 {
-  std::vector<WorldPluginPtr>::iterator iter;
-  for (iter = this->dataPtr->plugins.begin();
-      iter != this->dataPtr->plugins.end(); ++iter)
+  for (auto plugin = this->dataPtr->plugins.begin();
+           plugin != this->dataPtr->plugins.end(); ++plugin)
   {
-    if ((*iter)->GetHandle() == _name)
+    if ((*plugin)->GetHandle() == _name)
     {
-      this->dataPtr->plugins.erase(iter);
+      this->dataPtr->plugins.erase(plugin);
       break;
     }
   }
@@ -1321,11 +1315,9 @@ void World::ProcessEntityMsgs()
 {
   boost::mutex::scoped_lock lock(this->dataPtr->entityDeleteMutex);
 
-  std::list<std::string>::iterator iter;
-  for (iter = this->dataPtr->deleteEntity.begin();
-       iter != this->dataPtr->deleteEntity.end(); ++iter)
+  for (auto &entityName : this->dataPtr->deleteEntity)
   {
-    this->RemoveModel(*iter);
+    this->RemoveModel(entityName);
   }
 
   if (!this->dataPtr->deleteEntity.empty())
@@ -1341,16 +1333,14 @@ void World::ProcessRequestMsgs()
   boost::recursive_mutex::scoped_lock lock(*this->dataPtr->receiveMutex);
   msgs::Response response;
 
-  std::list<msgs::Request>::iterator iter;
-  for (iter = this->dataPtr->requestMsgs.begin();
-       iter != this->dataPtr->requestMsgs.end(); ++iter)
+  for (auto const &requestMsg : this->dataPtr->requestMsgs)
   {
     bool send = true;
-    response.set_id((*iter).id());
-    response.set_request((*iter).request());
+    response.set_id(requestMsg.id());
+    response.set_request(requestMsg.request());
     response.set_response("success");
 
-    if ((*iter).request() == "entity_list")
+    if (requestMsg.request() == "entity_list")
     {
       msgs::Model_V modelVMsg;
 
@@ -1370,14 +1360,14 @@ void World::ProcessRequestMsgs()
       std::string *serializedData = response.mutable_serialized_data();
       modelVMsg.SerializeToString(serializedData);
     }
-    else if ((*iter).request() == "entity_delete")
+    else if (requestMsg.request() == "entity_delete")
     {
       boost::mutex::scoped_lock lock2(this->dataPtr->entityDeleteMutex);
-      this->dataPtr->deleteEntity.push_back((*iter).data());
+      this->dataPtr->deleteEntity.push_back(requestMsg.data());
     }
-    else if ((*iter).request() == "entity_info")
+    else if (requestMsg.request() == "entity_info")
     {
-      BasePtr entity = this->dataPtr->rootElement->GetByName((*iter).data());
+      BasePtr entity = this->dataPtr->rootElement->GetByName(requestMsg.data());
       if (entity)
       {
         if (entity->HasType(Base::MODEL))
@@ -1428,7 +1418,7 @@ void World::ProcessRequestMsgs()
         response.set_response("nonexistant");
       }
     }
-    else if ((*iter).request() == "world_sdf")
+    else if (requestMsg.request() == "world_sdf")
     {
       msgs::GzString msg;
       this->UpdateStateSDF();
@@ -1444,7 +1434,7 @@ void World::ProcessRequestMsgs()
       msg.SerializeToString(serializedData);
       response.set_type(msg.GetTypeName());
     }
-    else if ((*iter).request() == "scene_info")
+    else if (requestMsg.request() == "scene_info")
     {
       this->dataPtr->sceneMsg.clear_model();
       this->BuildSceneMsg(this->dataPtr->sceneMsg, this->dataPtr->rootElement);
@@ -1453,7 +1443,7 @@ void World::ProcessRequestMsgs()
       this->dataPtr->sceneMsg.SerializeToString(serializedData);
       response.set_type(this->dataPtr->sceneMsg.GetTypeName());
     }
-    else if ((*iter).request() == "spherical_coordinates_info")
+    else if (requestMsg.request() == "spherical_coordinates_info")
     {
       msgs::SphericalCoordinates sphereCoordMsg;
       msgs::Set(&sphereCoordMsg, *(this->dataPtr->sphericalCoordinates));
@@ -1478,22 +1468,20 @@ void World::ProcessRequestMsgs()
 void World::ProcessModelMsgs()
 {
   boost::recursive_mutex::scoped_lock lock(*this->dataPtr->receiveMutex);
-  std::list<msgs::Model>::iterator iter;
-  for (iter = this->dataPtr->modelMsgs.begin();
-      iter != this->dataPtr->modelMsgs.end(); ++iter)
+  for (auto const &modelMsg : this->dataPtr->modelMsgs)
   {
     ModelPtr model;
-    if ((*iter).has_id())
-      model = this->GetModelById((*iter).id());
+    if (modelMsg.has_id())
+      model = this->GetModelById(modelMsg.id());
     else
-      model = this->GetModel((*iter).name());
+      model = this->GetModel(modelMsg.name());
 
     if (!model)
       gzerr << "Unable to find model["
-            << (*iter).name() << "] Id[" << (*iter).id() << "]\n";
+            << modelMsg.name() << "] Id[" << modelMsg.id() << "]\n";
     else
     {
-      model->ProcessMsg(*iter);
+      model->ProcessMsg(modelMsg);
 
       // May 30, 2013: The following code was removed because it has a
       // major performance impact when dragging complex object via the GUI.
@@ -1516,7 +1504,7 @@ void World::ProcessModelMsgs()
       //   }
       // }
 
-      this->dataPtr->modelPub->Publish(*iter);
+      this->dataPtr->modelPub->Publish(modelMsg);
     }
   }
 
@@ -1531,28 +1519,27 @@ void World::ProcessModelMsgs()
 void World::ProcessFactoryMsgs()
 {
   std::list<sdf::ElementPtr> modelsToLoad;
-  std::list<msgs::Factory>::iterator iter;
 
   {
     boost::recursive_mutex::scoped_lock lock(*this->dataPtr->receiveMutex);
-    for (iter = this->dataPtr->factoryMsgs.begin();
-        iter != this->dataPtr->factoryMsgs.end(); ++iter)
+    for (auto const &factoryMsg : this->dataPtr->factoryMsgs)
     {
       this->dataPtr->factorySDF->root->ClearElements();
 
-      if ((*iter).has_sdf() && !(*iter).sdf().empty())
+      if (factoryMsg.has_sdf() && !factoryMsg.sdf().empty())
       {
         // SDF Parsing happens here
-        if (!sdf::readString((*iter).sdf(), this->dataPtr->factorySDF))
+        if (!sdf::readString(factoryMsg.sdf(), this->dataPtr->factorySDF))
         {
-          gzerr << "Unable to read sdf string[" << (*iter).sdf() << "]\n";
+          gzerr << "Unable to read sdf string[" << factoryMsg.sdf() << "]\n";
           continue;
         }
       }
-      else if ((*iter).has_sdf_filename() && !(*iter).sdf_filename().empty())
+      else if (factoryMsg.has_sdf_filename() &&
+              !factoryMsg.sdf_filename().empty())
       {
         std::string filename = common::ModelDatabase::Instance()->GetModelFile(
-            (*iter).sdf_filename());
+            factoryMsg.sdf_filename());
 
         if (!sdf::readFile(filename, this->dataPtr->factorySDF))
         {
@@ -1560,12 +1547,12 @@ void World::ProcessFactoryMsgs()
           continue;
         }
       }
-      else if ((*iter).has_clone_model_name())
+      else if (factoryMsg.has_clone_model_name())
       {
-        ModelPtr model = this->GetModel((*iter).clone_model_name());
+        ModelPtr model = this->GetModel(factoryMsg.clone_model_name());
         if (!model)
         {
-          gzerr << "Unable to clone model[" << (*iter).clone_model_name()
+          gzerr << "Unable to clone model[" << factoryMsg.clone_model_name()
             << "]. Model not found.\n";
           continue;
         }
@@ -1592,10 +1579,10 @@ void World::ProcessFactoryMsgs()
         continue;
       }
 
-      if ((*iter).has_edit_name())
+      if (factoryMsg.has_edit_name())
       {
         BasePtr base =
-          this->dataPtr->rootElement->GetByName((*iter).edit_name());
+          this->dataPtr->rootElement->GetByName(factoryMsg.edit_name());
         if (base)
         {
           sdf::ElementPtr elem;
@@ -1649,8 +1636,8 @@ void World::ProcessFactoryMsgs()
 
         elem->SetParent(this->dataPtr->sdf);
         elem->GetParent()->InsertElement(elem);
-        if ((*iter).has_pose())
-          elem->GetElement("pose")->Set(msgs::Convert((*iter).pose()));
+        if (factoryMsg.has_pose())
+          elem->GetElement("pose")->Set(msgs::Convert(factoryMsg.pose()));
 
         if (isActor)
         {
@@ -1675,14 +1662,13 @@ void World::ProcessFactoryMsgs()
     this->dataPtr->factoryMsgs.clear();
   }
 
-  for (std::list<sdf::ElementPtr>::iterator iter2 = modelsToLoad.begin();
-       iter2 != modelsToLoad.end(); ++iter2)
+  for (auto const &elem : modelsToLoad)
   {
     try
     {
       boost::mutex::scoped_lock lock(this->dataPtr->factoryDeleteMutex);
 
-      ModelPtr model = this->LoadModel(*iter2, this->dataPtr->rootElement);
+      ModelPtr model = this->LoadModel(elem, this->dataPtr->rootElement);
       model->Init();
       model->LoadPlugins();
     }
@@ -1730,14 +1716,13 @@ void World::SetState(const WorldState &_state)
   this->dataPtr->logRealTime = _state.GetRealTime();
 
   const ModelState_M modelStates = _state.GetModelStates();
-  for (ModelState_M::const_iterator iter = modelStates.begin();
-       iter != modelStates.end(); ++iter)
+  for (auto const &modelState : modelStates)
   {
-    ModelPtr model = this->GetModel(iter->second.GetName());
+    ModelPtr model = this->GetModel(modelState.second.GetName());
     if (model)
-      model->SetState(iter->second);
+      model->SetState(modelState.second);
     else
-      gzerr << "Unable to find model[" << iter->second.GetName() << "]\n";
+      gzerr << "Unable to find model[" << modelState.second.GetName() << "]\n";
   }
 }
 
@@ -1780,20 +1765,18 @@ std::string World::StripWorldName(const std::string &_name) const
 //////////////////////////////////////////////////
 void World::EnableAllModels()
 {
-  for (Model_V::iterator iter = this->dataPtr->models.begin();
-       iter != this->dataPtr->models.end(); ++iter)
+  for (auto &model : this->dataPtr->models)
   {
-    (*iter)->SetEnabled(true);
+    model->SetEnabled(true);
   }
 }
 
 //////////////////////////////////////////////////
 void World::DisableAllModels()
 {
-  for (Model_V::iterator iter = this->dataPtr->models.begin();
-       iter != this->dataPtr->models.end(); ++iter)
+  for (auto &model : this->dataPtr->models)
   {
-    (*iter)->SetEnabled(false);
+    model->SetEnabled(false);
   }
 }
 
@@ -1828,11 +1811,11 @@ bool World::OnLog(std::ostringstream &_stream)
       boost::mutex::scoped_lock lock(this->dataPtr->logBufferMutex);
       this->dataPtr->currentStateBuffer ^= 1;
     }
-    for (std::deque<WorldState>::iterator iter =
-        this->dataPtr->states[bufferIndex].begin();
-        iter != this->dataPtr->states[bufferIndex].end(); ++iter)
+    for (auto const &worldState : this->dataPtr->states[bufferIndex])
     {
-      _stream << "<sdf version='" << SDF_VERSION << "'>" << *iter << "</sdf>";
+      _stream << "<sdf version='" << SDF_VERSION << "'>"
+              << worldState
+              << "</sdf>";
     }
 
     this->dataPtr->states[bufferIndex].clear();
@@ -1891,26 +1874,23 @@ void World::ProcessMessages()
 
       if (!this->dataPtr->publishModelPoses.empty())
       {
-        for (std::set<ModelPtr>::iterator iter =
-            this->dataPtr->publishModelPoses.begin();
-            iter != this->dataPtr->publishModelPoses.end(); ++iter)
+        for (auto const &model : this->dataPtr->publishModelPoses)
         {
           msgs::Pose *poseMsg = msg.add_pose();
 
           // Publish the model's relative pose
-          poseMsg->set_name((*iter)->GetScopedName());
-          poseMsg->set_id((*iter)->GetId());
-          msgs::Set(poseMsg, (*iter)->GetRelativePose());
+          poseMsg->set_name(model->GetScopedName());
+          poseMsg->set_id(model->GetId());
+          msgs::Set(poseMsg, model->GetRelativePose());
 
           // Publish each of the model's children relative poses
-          Link_V links = (*iter)->GetLinks();
-          for (Link_V::iterator linkIter = links.begin();
-              linkIter != links.end(); ++linkIter)
+          Link_V links = model->GetLinks();
+          for (auto const &link : links)
           {
             poseMsg = msg.add_pose();
-            poseMsg->set_name((*linkIter)->GetScopedName());
-            poseMsg->set_id((*linkIter)->GetId());
-            msgs::Set(poseMsg, (*linkIter)->GetRelativePose());
+            poseMsg->set_name(link->GetScopedName());
+            poseMsg->set_id(link->GetId());
+            msgs::Set(poseMsg, link->GetRelativePose());
           }
         }
 
@@ -2033,16 +2013,16 @@ void World::RemoveModel(const std::string &_name)
 
   // Remove all the dirty poses from the delete entity.
   {
-    for (std::list<Entity*>::iterator iter2 = this->dataPtr->dirtyPoses.begin();
-        iter2 != this->dataPtr->dirtyPoses.end();)
+    for (auto entity = this->dataPtr->dirtyPoses.begin();
+             entity != this->dataPtr->dirtyPoses.end(); ++entity)
     {
-      if ((*iter2)->GetName() == _name ||
-          ((*iter2)->GetParent() && (*iter2)->GetParent()->GetName() == _name))
+      if ((*entity)->GetName() == _name ||
+         ((*entity)->GetParent() && (*entity)->GetParent()->GetName() == _name))
       {
-        this->dataPtr->dirtyPoses.erase(iter2++);
+        this->dataPtr->dirtyPoses.erase(entity++);
       }
       else
-        ++iter2;
+        ++entity;
     }
   }
 
@@ -2083,12 +2063,12 @@ void World::RemoveModel(const std::string &_name)
 
     this->dataPtr->rootElement->RemoveChild(_name);
 
-    for (Model_V::iterator iter = this->dataPtr->models.begin();
-        iter != this->dataPtr->models.end(); ++iter)
+    for (auto model = this->dataPtr->models.begin();
+             model != this->dataPtr->models.end(); ++model)
     {
-      if ((*iter)->GetName() == _name || (*iter)->GetScopedName() == _name)
+      if ((*model)->GetName() == _name || (*model)->GetScopedName() == _name)
       {
-        this->dataPtr->models.erase(iter);
+        this->dataPtr->models.erase(model);
         break;
       }
     }
@@ -2097,13 +2077,12 @@ void World::RemoveModel(const std::string &_name)
   // Cleanup the publishModelPoses list.
   {
     boost::recursive_mutex::scoped_lock lock2(*this->dataPtr->receiveMutex);
-    for (std::set<ModelPtr>::iterator iter =
-         this->dataPtr->publishModelPoses.begin();
-         iter != this->dataPtr->publishModelPoses.end(); ++iter)
+    for (auto model = this->dataPtr->publishModelPoses.begin();
+             model != this->dataPtr->publishModelPoses.end(); ++model)
     {
-      if ((*iter)->GetName() == _name || (*iter)->GetScopedName() == _name)
+      if ((*model)->GetName() == _name || (*model)->GetScopedName() == _name)
       {
-        this->dataPtr->publishModelPoses.erase(iter);
+        this->dataPtr->publishModelPoses.erase(model);
         break;
       }
     }